refactor: share github remote parsing
